requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
Django==3.0.5
django-filter==2.2.0
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
